Går over til å bruke ny måte for å sjekke om berørt behandling skal opprettes
