+++ domenetjenester/behandling-revurdering/src/main/java/no/nav/foreldrepenger/behandling/revurdering/BerørtBehandlingTjeneste.java
@@ -9,9 +9,6 @@ import java.util.stream.Collectors;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import no.nav.foreldrepenger.behandling.revurdering.ytelse.UttakInputTjeneste;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandlingsresultat;
@@ -24,20 +21,16 @@ import no.nav.foreldrepenger.behandlingslager.behandling.historikk.HistorikkRepo
 import no.nav.foreldrepenger.behandlingslager.behandling.historikk.Historikkinnslag;
 import no.nav.foreldrepenger.behandlingslager.behandling.historikk.HistorikkinnslagType;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
-import no.nav.foreldrepenger.behandlingslager.uttak.PeriodeResultatType;
 import no.nav.foreldrepenger.domene.uttak.ForeldrepengerUttak;
 import no.nav.foreldrepenger.domene.uttak.ForeldrepengerUttakPeriode;
 import no.nav.foreldrepenger.domene.uttak.ForeldrepengerUttakTjeneste;
 import no.nav.foreldrepenger.domene.uttak.input.ForeldrepengerGrunnlag;
-import no.nav.foreldrepenger.domene.uttak.input.UttakInput;
 import no.nav.foreldrepenger.domene.uttak.saldo.StønadskontoSaldoTjeneste;
 import no.nav.foreldrepenger.domene.ytelsefordeling.YtelseFordelingTjeneste;
 import no.nav.foreldrepenger.historikk.HistorikkInnslagTekstBuilder;
 
 @ApplicationScoped
 public class BerørtBehandlingTjeneste {
-
-    private static final Logger LOG = LoggerFactory.getLogger(BerørtBehandlingTjeneste.class);
 
     private StønadskontoSaldoTjeneste stønadskontoSaldoTjeneste;
     private HistorikkRepository historikkRepository;
@@ -71,41 +64,25 @@ public class BerørtBehandlingTjeneste {
      *                                            behandling.
      * @return true dersom berørt behandling skal opprettes, ellers false.
      */
-    public boolean skalBerørtBehandlingOpprettes(Optional<Behandlingsresultat> brukersGjeldendeBehandlingsresultat,
+    public boolean skalBerørtBehandlingOpprettes(Behandlingsresultat brukersGjeldendeBehandlingsresultat,
                                                  Long behandlingId,
                                                  Long behandlingIdAnnenPart) {
-        var gammel = gammel(brukersGjeldendeBehandlingsresultat, behandlingId, behandlingIdAnnenPart);
-        try {
-            var ny = skalBerørtBehandlingOpprettesNy(brukersGjeldendeBehandlingsresultat.orElseThrow(), behandlingId,
-                behandlingIdAnnenPart);
-            if (gammel != ny) {
-                LOG.info("Ny utregning av berørt behandling gammel:{}, ny:{}", gammel, ny);
-            }
-
-        } catch (Throwable t) {
-            LOG.info("Ny utregning av berørt behandling feilet", t);
-        }
-        return gammel;
-    }
-
-    boolean skalBerørtBehandlingOpprettesNy(Behandlingsresultat behandlingsresultat,
-                                            Long behandlingId,
-                                            Long behandlingIdAnnenpart) {
         //Må sjekke konsekvens pga overlapp med samtidig uttak
-        if (behandlingsresultat.isBehandlingHenlagt() || harKonsekvens(behandlingsresultat, KonsekvensForYtelsen.INGEN_ENDRING)) {
+        if (brukersGjeldendeBehandlingsresultat.isBehandlingHenlagt() || harKonsekvens(
+            brukersGjeldendeBehandlingsresultat, KonsekvensForYtelsen.INGEN_ENDRING)) {
             return false;
         }
-        var uttakInput = uttakInputTjeneste.lagInput(behandlingsresultat.getBehandlingId());
+        var uttakInput = uttakInputTjeneste.lagInput(brukersGjeldendeBehandlingsresultat.getBehandlingId());
         ForeldrepengerGrunnlag foreldrepengerGrunnlag = uttakInput.getYtelsespesifiktGrunnlag();
         if (foreldrepengerGrunnlag.isTapendeBehandling()) {
             return false;
         }
 
-        if (behandlingsresultat.isEndretStønadskonto() || stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)) {
+        if (brukersGjeldendeBehandlingsresultat.isEndretStønadskonto() || stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)) {
             return true;
         }
 
-        var annenpartsUttak = hentUttak(behandlingIdAnnenpart);
+        var annenpartsUttak = hentUttak(behandlingIdAnnenPart);
         if (annenpartsUttak.isEmpty()) {
             return false;
         }
@@ -174,162 +151,8 @@ public class BerørtBehandlingTjeneste {
         return uttak.getGjeldendePerioder().stream().anyMatch(p -> p.getTom().isAfter(dato));
     }
 
-    private boolean gammel(Optional<Behandlingsresultat> brukersGjeldendeBehandlingsresultat,
-                           Long behandlingId,
-                           Long behandlingIdAnnenPart) {
-        if (brukersGjeldendeBehandlingsresultat.isEmpty() || (behandlingId == null)) {
-            return false;
-        }
-        var uttakInput = uttakInputTjeneste.lagInput(brukersGjeldendeBehandlingsresultat.get().getBehandlingId());
-        ForeldrepengerGrunnlag foreldrepengerGrunnlag = uttakInput.getYtelsespesifiktGrunnlag();
-        if ((foreldrepengerGrunnlag == null) || foreldrepengerGrunnlag.isTapendeBehandling()) {
-            return false;
-        }
-        var brukersUttaksPerioder = hentUttak(behandlingId);
-        var medforeldersUttaksPerioder = hentUttak(behandlingIdAnnenPart);
-
-        return skalBerørtBehandlingOpprettes(brukersGjeldendeBehandlingsresultat, brukersUttaksPerioder,
-            medforeldersUttaksPerioder);
-    }
-
-    /**
-     * Finner ut om det skal opprettes en berørt behandling på med forelders sak.
-     *
-     * @param optionalBehandlingsresultat brukers behandlingsresultat
-     * @param brukersUttaksPerioder       brukers uttaksplan.
-     * @param medforeldersUttaksPerioder  medforelders uttaksplan.
-     * @return true dersom berørt behandling skal opprettes, ellers false.
-     */
-    boolean skalBerørtBehandlingOpprettes(Optional<Behandlingsresultat> optionalBehandlingsresultat,
-                                          Optional<ForeldrepengerUttak> brukersUttaksPerioder,
-                                          Optional<ForeldrepengerUttak> medforeldersUttaksPerioder) {
-        if (optionalBehandlingsresultat.isEmpty()) {
-            return false;
-        }
-        Behandlingsresultat behandlingsresultat = optionalBehandlingsresultat.get();
-        var uttakInput = uttakInputTjeneste.lagInput(behandlingsresultat.getBehandlingId());
-        return skalRevurderingFøreTilBerørtBehandling(behandlingsresultat, brukersUttaksPerioder,
-            medforeldersUttaksPerioder, uttakInput);
-    }
-
     private Optional<ForeldrepengerUttak> hentUttak(Long behandling) {
         return uttakTjeneste.hentUttakHvisEksisterer(behandling);
-    }
-
-    private boolean skalRevurderingFøreTilBerørtBehandling(Behandlingsresultat behandlingsresultat,
-                                                           Optional<ForeldrepengerUttak> brukersUttaksPerioder,
-                                                           Optional<ForeldrepengerUttak> medforeldersUttaksPerioder,
-                                                           UttakInput uttakInput) {
-        if (behandlingsresultat.isBehandlingsresultatOpphørt() && harMedforelderPerioderEtterBrukersOpphør(
-            brukersUttaksPerioder, medforeldersUttaksPerioder)) {
-            var harKonsekvens = harKonsekvens(behandlingsresultat, KonsekvensForYtelsen.FORELDREPENGER_OPPHØRER);
-            if (harKonsekvens) {
-                log("OPPHØR");
-            }
-            return harKonsekvens;
-        }
-        if (behandlingsresultat.isBehandlingsresultatForeldrepengerEndret()
-            || behandlingsresultat.isBehandlingsresultatInnvilget()) {
-            if (harKonsekvens(behandlingsresultat, KonsekvensForYtelsen.ENDRING_I_BEREGNING) || harKonsekvens(
-                behandlingsresultat, KonsekvensForYtelsen.ENDRING_I_UTTAK)) {
-                if (behandlingsresultat.isEndretStønadskonto()) {
-                    log("ENDRET_KONTO");
-                    return true;
-                }
-            }
-            if (stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)) {
-                log("NEGATIV_KONTO");
-                return true;
-            }
-            var harOverlappendePerioder = harOverlappendePerioder(brukersUttaksPerioder, medforeldersUttaksPerioder);
-            if (harOverlappendePerioder) {
-                log("OVERLAPP_UTTAK");
-            }
-            return harOverlappendePerioder;
-        }
-        return false;
-    }
-
-    private void log(String årsak) {
-        LOG.info("Berørt behandling opprettes pga {}", årsak);
-    }
-
-    private boolean harMedforelderPerioderEtterBrukersOpphør(Optional<ForeldrepengerUttak> brukersUttaksplan,
-                                                             Optional<ForeldrepengerUttak> motpartsUttaksplan) {
-        if (brukersUttaksplan.isPresent() && motpartsUttaksplan.isPresent()) {
-            if (harBrukerFørsteUttak(brukersUttaksplan.get(), motpartsUttaksplan.get())) {
-                return true;
-            }
-            Optional<LocalDate> medforeldrersSisteDag = sisteUttaksdato(motpartsUttaksplan.get());
-            Optional<LocalDate> brukersFørsteDag = førsteUttaksdatoUtenAvslåttePerioder(brukersUttaksplan.get());
-            if (medforeldrersSisteDag.isPresent() && brukersFørsteDag.isPresent()) {
-                return !brukersFørsteDag.get().isAfter(medforeldrersSisteDag.get());
-            }
-        }
-        return false;
-    }
-
-    private boolean harBrukerFørsteUttak(ForeldrepengerUttak brukersUttaksplan,
-                                         ForeldrepengerUttak motpartsUttaksplan) {
-        return brukersUttaksplan.getGjeldendePerioder()
-            .stream()
-            .map(p -> p.getFom())
-            .min(LocalDate::compareTo)
-            .get()
-            .isBefore(motpartsUttaksplan.getGjeldendePerioder()
-                .stream()
-                .map(p -> p.getFom())
-                .min(LocalDate::compareTo)
-                .get());
-    }
-
-    private boolean harOverlappendePerioder(Optional<ForeldrepengerUttak> brukersUttaksplan,
-                                            Optional<ForeldrepengerUttak> motpartsUttaksplan) {
-        if (brukersUttaksplan.isPresent() && motpartsUttaksplan.isPresent()) {
-            LocalDate førsteForeldrersSisteDag = førsteForeldrersSisteDag(brukersUttaksplan.get(),
-                motpartsUttaksplan.get());
-            LocalDate andreForeldrersFørsteDag = andreForeldrersFørsteDag(brukersUttaksplan.get(),
-                motpartsUttaksplan.get());
-            if ((førsteForeldrersSisteDag != null) && (andreForeldrersFørsteDag != null)) {
-                return !andreForeldrersFørsteDag.isAfter(førsteForeldrersSisteDag);
-            }
-        }
-        return false;
-    }
-
-    private LocalDate førsteForeldrersSisteDag(ForeldrepengerUttak brukersUttaksplan,
-                                               ForeldrepengerUttak motpartsUttaksplan) {
-        Optional<LocalDate> førsteForeldrersSisteDag = brukersUttaksplan.finnFørsteUttaksdato()
-            .isAfter(motpartsUttaksplan.finnFørsteUttaksdato()) ? sisteUttaksdato(motpartsUttaksplan) : sisteUttaksdato(
-            brukersUttaksplan);
-        return førsteForeldrersSisteDag.orElse(null);
-    }
-
-    private LocalDate andreForeldrersFørsteDag(ForeldrepengerUttak brukersUttaksplan,
-                                               ForeldrepengerUttak motpartsUttaksplan) {
-        Optional<LocalDate> andreForeldrersFørsteDag = brukersUttaksplan.finnFørsteUttaksdato()
-            .isAfter(motpartsUttaksplan.finnFørsteUttaksdato()) ? førsteUttaksdatoUtenAvslåttePerioder(
-            brukersUttaksplan) : førsteUttaksdatoUtenAvslåttePerioder(motpartsUttaksplan);
-        return andreForeldrersFørsteDag.orElse(null);
-    }
-
-    private Optional<LocalDate> førsteUttaksdatoUtenAvslåttePerioder(ForeldrepengerUttak uttaksplan) {
-        return uttaksplan.getGjeldendePerioder()
-            .stream()
-            .filter(u -> !PeriodeResultatType.AVSLÅTT.equals(u.getResultatType()))
-            .map(p -> p.getFom())
-            .min(LocalDate::compareTo);
-    }
-
-    private Optional<LocalDate> sisteUttaksdato(ForeldrepengerUttak uttaksplan) {
-        return uttaksplan.getGjeldendePerioder()
-            .stream()
-            .filter(
-                periode -> PeriodeResultatType.INNVILGET.equals(periode.getResultatType()) || periode.getAktiviteter()
-                    .stream()
-                    .anyMatch(a -> a.getTrekkdager().merEnn0()))
-            .map(p -> p.getTidsperiode().getTomDato())
-            .max(LocalDate::compareTo);
     }
 
     public boolean harKonsekvens(Behandlingsresultat behandlingsresultat, KonsekvensForYtelsen konsekvens) {
