+++ domenetjenester/behandling-revurdering/src/test/java/no/nav/foreldrepenger/behandling/revurdering/BerørtBehandlingTjenesteTest.java
@@ -1,39 +1,42 @@
 package no.nav.foreldrepenger.behandling.revurdering;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import javax.persistence.EntityManager;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 import no.nav.foreldrepenger.behandling.BehandlingReferanse;
 import no.nav.foreldrepenger.behandling.revurdering.ytelse.UttakInputTjeneste;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingResultatType;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingType;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandlingsresultat;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingÅrsak;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingÅrsakType;
+import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingsresultatRepository;
 import no.nav.foreldrepenger.behandlingslager.behandling.KonsekvensForYtelsen;
-import no.nav.foreldrepenger.behandlingslager.behandling.historikk.HistorikkRepository;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
-import no.nav.foreldrepenger.behandlingslager.fagsak.Fagsak;
-import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
-import no.nav.foreldrepenger.behandlingslager.testutilities.aktør.NavBrukerBuilder;
+import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.AvklarteUttakDatoerEntitet;
+import no.nav.foreldrepenger.behandlingslager.testutilities.behandling.ScenarioFarSøkerForeldrepenger;
+import no.nav.foreldrepenger.behandlingslager.testutilities.behandling.ScenarioMorSøkerForeldrepenger;
 import no.nav.foreldrepenger.behandlingslager.uttak.PeriodeResultatType;
 import no.nav.foreldrepenger.behandlingslager.uttak.UttakArbeidType;
-import no.nav.foreldrepenger.behandlingslager.uttak.fp.FpUttakRepository;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.IkkeOppfyltÅrsak;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.InnvilgetÅrsak;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.PeriodeResultatÅrsak;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.SamtidigUttaksprosent;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.StønadskontoType;
 import no.nav.foreldrepenger.behandlingslager.uttak.fp.Trekkdager;
-import no.nav.foreldrepenger.domene.typer.AktørId;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.Utbetalingsgrad;
+import no.nav.foreldrepenger.behandlingslager.uttak.fp.UttakUtsettelseType;
+import no.nav.foreldrepenger.dbstoette.FPsakEntityManagerAwareExtension;
 import no.nav.foreldrepenger.domene.uttak.ForeldrepengerUttak;
 import no.nav.foreldrepenger.domene.uttak.ForeldrepengerUttakAktivitet;
 import no.nav.foreldrepenger.domene.uttak.ForeldrepengerUttakPeriode;
@@ -44,232 +47,518 @@ import no.nav.foreldrepenger.domene.uttak.input.UttakInput;
 import no.nav.foreldrepenger.domene.uttak.saldo.StønadskontoSaldoTjeneste;
 import no.nav.foreldrepenger.domene.ytelsefordeling.YtelseFordelingTjeneste;
 
-@ExtendWith(MockitoExtension.class)
-public class BerørtBehandlingTjenesteTest {
+@ExtendWith(FPsakEntityManagerAwareExtension.class)
+class BerørtBehandlingTjenesteTest {
 
-    private BerørtBehandlingTjeneste berørtBehandlingTjeneste;
-    @Mock
-    private StønadskontoSaldoTjeneste stønadskontoSaldoTjeneste;
-    @Mock
-    private HistorikkRepository historikkRepository;
-    @Mock
     private BehandlingRepositoryProvider repositoryProvider;
-    @Mock
-    private FpUttakRepository fpUttakRepository;
-    @Mock
-    private UttakInputTjeneste uttakInputTjeneste;
-    @Mock
-    private YtelseFordelingTjeneste ytelseFordelingTjeneste;
+    private BerørtBehandlingTjeneste tjeneste;
 
-    private Behandling revurdering;
+    private ForeldrepengerUttakTjeneste foreldrepengerUttakTjeneste;
+    private UttakInputTjeneste uttakInputTjeneste;
+    private StønadskontoSaldoTjeneste saldoTjeneste;
 
     @BeforeEach
-    public void fellesOppsett() {
-        when(repositoryProvider.getHistorikkRepository()).thenReturn(historikkRepository);
-        when(repositoryProvider.getFpUttakRepository()).thenReturn(fpUttakRepository);
-        var fagsak = Fagsak.opprettNy(FagsakYtelseType.FORELDREPENGER,
-                new NavBrukerBuilder().medAktørId(AktørId.dummy()).build());
-        var førsteBehandling = Behandling.forFørstegangssøknad(fagsak).build();
-        revurdering = Behandling.fraTidligereBehandling(førsteBehandling, BehandlingType.REVURDERING).build();
-        berørtBehandlingTjeneste = new BerørtBehandlingTjeneste(stønadskontoSaldoTjeneste, repositoryProvider,
-                uttakInputTjeneste,
-                new ForeldrepengerUttakTjeneste(repositoryProvider.getFpUttakRepository()), ytelseFordelingTjeneste);
+    void setUp(EntityManager entityManager) {
+        repositoryProvider = new BehandlingRepositoryProvider(entityManager);
+        foreldrepengerUttakTjeneste = mock(ForeldrepengerUttakTjeneste.class);
+        uttakInputTjeneste = mock(UttakInputTjeneste.class);
+        saldoTjeneste = mock(StønadskontoSaldoTjeneste.class);
+        tjeneste = new BerørtBehandlingTjeneste(saldoTjeneste, repositoryProvider, uttakInputTjeneste,
+            foreldrepengerUttakTjeneste,
+            new YtelseFordelingTjeneste(repositoryProvider.getYtelsesFordelingRepository()));
     }
 
-    // Scenarie 1 - Opphør
     @Test
-    public void skal_opprette_berørt_behandling_dersom_behandlingsresultater_er_opphør() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.OPPHØR,
-                KonsekvensForYtelsen.FORELDREPENGER_OPPHØRER, false);
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(revurdering), null, new ForeldrepengerGrunnlag());
-        when(uttakInputTjeneste.lagInput(revurdering.getId())).thenReturn(uttakInput);
+    public void tapende_behandling_skal_ikke_opprette_berørt() {
+        var scenario = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var scenarioAnnenpart = ScenarioFarSøkerForeldrepenger.forFødsel();
+        var behandling = scenario.lagre(repositoryProvider);
+        var behandlingAnnenpart = scenarioAnnenpart.lagre(repositoryProvider);
 
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat,
-                lagUttakResultatPerioder(LocalDate.now().minusDays(5), LocalDate.now().plusDays(5), false, false),
-                lagUttakResultatPerioder(LocalDate.now().plusDays(6), LocalDate.now().plusDays(12), false,
-                        false))).isTrue();
-    }
-
-    // Scenarie 2a - endring i beregning
-    @Test
-    public void skal_opprette_berørt_behandling_dersom_revudering_er_innvilget_med_endring_i_stønadskonto() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.INNVILGET,
-                KonsekvensForYtelsen.ENDRING_I_BEREGNING, true);
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(revurdering), null, new ForeldrepengerGrunnlag());
-        when(uttakInputTjeneste.lagInput(revurdering.getId())).thenReturn(uttakInput);
-
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat, Optional.empty(),
-                Optional.empty())).isTrue();
-    }
-
-    // Scenarie 2b - endring i uttak
-    @Test
-    public void skal_opprette_berørt_behandling_dersom_revudering_er_innvilget_med_endring_i_uttak() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.INNVILGET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, true);
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(revurdering), null, new ForeldrepengerGrunnlag());
-        when(uttakInputTjeneste.lagInput(revurdering.getId())).thenReturn(uttakInput);
-
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat, Optional.empty(),
-                Optional.empty())).isTrue();
-    }
-
-    // Scenarie 4a - foreldrepenger er endret og uttak har overlappende perioder
-    @Test
-    public void skal_opprette_berørt_behandling_dersom_revudering_er_endring_i_foreldrepenger_med_endring_i_uttak_overlappende_perioder() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.FORELDREPENGER_ENDRET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, false);
-
-        var behandling = revurdering;
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(behandling), null, new ForeldrepengerGrunnlag());
-        when(stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(false);
+        var uttakInput = new UttakInput(BehandlingReferanse.fra(behandling), null,
+            new ForeldrepengerGrunnlag().medErTapendeBehandling(true));
         when(uttakInputTjeneste.lagInput(behandling.getId())).thenReturn(uttakInput);
 
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat,
-                lagUttakResultatPerioder(LocalDate.now().minusDays(5), LocalDate.now().plusDays(5), false, false),
-                lagUttakResultatPerioder(LocalDate.now().minusDays(12), LocalDate.now().plusDays(2), false,
-                        false))).isTrue();
+        var resultat = skalBerørtOpprettes(behandling, behandlingAnnenpart);
+
+        assertThat(resultat).isFalse();
     }
 
-    // foreldrepenger er endret og uttak har ikke overlappende perioder
     @Test
-    public void skal_ikke_opprette_berørt_behandling_dersom_revudering_er_endring_i_foreldrepenger_med_endring_i_uttak_og_ikke_overlappende_perioder() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.FORELDREPENGER_ENDRET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, false);
+    public void annenpart_uten_uttak_skal_ikke_opprette_berørt() {
+        var scenario = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var scenarioAnnenpart = ScenarioFarSøkerForeldrepenger.forFødsel();
+        var behandling = scenario.lagre(repositoryProvider);
+        var behandlingAnnenpart = scenarioAnnenpart.lagre(repositoryProvider);
 
-        var behandling = revurdering;
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(behandling), null, new ForeldrepengerGrunnlag());
-        when(stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(false);
-        when(uttakInputTjeneste.lagInput(behandling.getId())).thenReturn(uttakInput);
+        lagUttakInput(behandling);
 
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat,
-                lagUttakResultatPerioder(LocalDate.now().minusDays(12), LocalDate.now().plusDays(2), true, false),
-                lagUttakResultatPerioder(LocalDate.now().plusDays(3), LocalDate.now().plusDays(5), false,
-                        false))).isFalse();
+        lagreUttak(behandlingAnnenpart, null);
+
+        var resultat = skalBerørtOpprettes(behandling, behandlingAnnenpart);
+
+        assertThat(resultat).isFalse();
     }
 
-    // foreldrepenger er endret og uttak har overlappende avslått periode
     @Test
-    public void skal_ikke_opprette_berørt_behandling_dersom_revudering_er_endring_i_foreldrepenger_med_endring_i_uttak_og_overlappende_avslått_perioder() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.FORELDREPENGER_ENDRET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, false);
+    public void behandling_uten_uttak_skal_ikke_opprette_berørt() {
+        var scenario = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var scenarioAnnenpart = ScenarioFarSøkerForeldrepenger.forFødsel();
+        var behandling = scenario.lagre(repositoryProvider);
+        var behandlingAnnenpart = scenarioAnnenpart.lagre(repositoryProvider);
 
-        var behandling = revurdering;
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(behandling), null, new ForeldrepengerGrunnlag());
-        when(uttakInputTjeneste.lagInput(behandling.getId())).thenReturn(uttakInput);
-        when(stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(false);
+        lagUttakInput(behandling);
 
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat,
-                lagUttakResultatPerioder(LocalDate.now().minusDays(12), LocalDate.now().plusDays(2), false, true),
-                lagUttakResultatPerioder(LocalDate.now().plusDays(3), LocalDate.now().plusDays(5), false,
-                        false))).isFalse();
+        var annenpartsUttak = new ForeldrepengerUttak(List.of(
+            new ForeldrepengerUttakPeriode.Builder().medTidsperiode(LocalDate.now(), LocalDate.now().plusWeeks(10))
+                .medResultatÅrsak(InnvilgetÅrsak.FELLESPERIODE_ELLER_FORELDREPENGER)
+                .build()));
+        lagreUttak(behandlingAnnenpart, annenpartsUttak);
+        lagreUttak(behandling, null);
+
+        var resultat = skalBerørtOpprettes(behandling, behandlingAnnenpart);
+
+        assertThat(resultat).isFalse();
     }
 
-    // foreldrepenger er endret og uttak har overlappende avslått periode med
-    // trekkdager
     @Test
-    public void skal_opprette_berørt_behandling_dersom_revudering_er_endring_i_foreldrepenger_med_endring_i_uttak_og_overlappende_avslått_perioder_med_trekkdager() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.FORELDREPENGER_ENDRET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, false);
+    public void behandling_med_negativ_saldo_skal_opprette_berørt() {
+        var scenario = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var scenarioAnnenpart = ScenarioFarSøkerForeldrepenger.forFødsel();
+        var behandling = scenario.lagre(repositoryProvider);
+        var behandlingAnnenpart = scenarioAnnenpart.lagre(repositoryProvider);
 
-        var behandling = revurdering;
+        var uttakInput = lagUttakInput(behandling);
+
+        when(saldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(true);
+
+        var resultat = skalBerørtOpprettes(behandling, behandlingAnnenpart);
+
+        assertThat(resultat).isTrue();
+    }
+
+    private UttakInput lagUttakInput(Behandling behandling) {
         var uttakInput = new UttakInput(BehandlingReferanse.fra(behandling), null, new ForeldrepengerGrunnlag());
         when(uttakInputTjeneste.lagInput(behandling.getId())).thenReturn(uttakInput);
-        when(stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(false);
+        return uttakInput;
+    }
 
-        var brukersPeriode = new ForeldrepengerUttakPeriode.Builder()
-                .medTidsperiode(LocalDate.of(2020, 1, 1), LocalDate.of(2020, 1, 1).plusWeeks(1))
-                .medResultatType(PeriodeResultatType.INNVILGET)
-                .build();
-        List<ForeldrepengerUttakPeriode> brukersPerioder = List.of(brukersPeriode);
-        var brukersUttak = new ForeldrepengerUttak(brukersPerioder);
-        var annenpartInnvilgetPeriode = new ForeldrepengerUttakPeriode.Builder()
-                .medTidsperiode(brukersPeriode.getFom().minusWeeks(1), brukersPeriode.getFom().minusDays(1))
-                .medResultatType(PeriodeResultatType.INNVILGET)
-                .build();
-        var aktivitetMedTrekkdager = new ForeldrepengerUttakPeriodeAktivitet.Builder()
-                .medTrekkdager(new Trekkdager(10))
-                .medArbeidsprosent(BigDecimal.ZERO)
+    @Test
+    public void behandling_med_endret_stønadskonto_skal_opprette_berørt() {
+        var scenario = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var scenarioAnnenpart = ScenarioFarSøkerForeldrepenger.forFødsel();
+        var behandling = scenario.lagre(repositoryProvider);
+        var behandlingAnnenpart = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(behandling);
+
+        var br = Behandlingsresultat.builder().medEndretStønadskonto(true).buildFor(behandling);
+
+        var resultat = tjeneste.skalBerørtBehandlingOpprettes(br, behandling.getId(), behandlingAnnenpart.getId());
+
+        assertThat(resultat).isTrue();
+    }
+
+    @Test
+    public void ikke_berørt_behandling_når_far_har_uttak_i_mellom_mors_uttaksperioder() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+        var startDatoFar = startDatoMor.plusMonths(3);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoFar));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoFar.minusDays(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoFar,
+            startDatoFar.plusWeeks(10))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morsAndrePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(farsPeriode.getTom().plusDays(1),
+            farsPeriode.getTom().plusWeeks(5))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsFørstePeriode, morsAndrePeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
+    }
+
+    @Test
+    public void ikke_berørt_behandling_når_far_har_overlapp_med_oppholdsperiode_i_mors_uttak() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+        var startDatoFar = startDatoMor.plusMonths(3);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoFar));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoFar.minusDays(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var morsOppholdsperiode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(
+            morsFørstePeriode.getTom().plusDays(1), startDatoFar).medResultatÅrsak(PeriodeResultatÅrsak.UKJENT).build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsOppholdsperiode.getFom(),
+            morsOppholdsperiode.getTom())
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morsAndrePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(farsPeriode.getTom().plusDays(1),
+            farsPeriode.getTom().plusWeeks(5))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsFørstePeriode, morsOppholdsperiode, morsAndrePeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
+    }
+
+    @Test
+    public void berørt_behandling_ved_overlapp_mellom_partene() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(15))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsPeriode.getFom().plusWeeks(1),
+            morsPeriode.getTom().minusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsPeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isTrue();
+    }
+
+    @Test
+    public void berørt_behandling_ved_overlapp_mellom_partene_der_annenpart_har_innvilget_utsettelse() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(15))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medResultatType(PeriodeResultatType.INNVILGET)
+            .medUtsettelseType(UttakUtsettelseType.FERIE)
+            .medAktiviteter(List.of(new ForeldrepengerUttakPeriodeAktivitet.Builder().medArbeidsprosent(BigDecimal.ZERO)
+                .medTrekkdager(Trekkdager.ZERO)
+                .medUtbetalingsgrad(Utbetalingsgrad.ZERO)
                 .medAktivitet(new ForeldrepengerUttakAktivitet(UttakArbeidType.FRILANS))
-                .build();
-        var annenpartOverlappendeAvslåttMedTrekkdager = new ForeldrepengerUttakPeriode.Builder()
-                .medTidsperiode(brukersPeriode.getFom(), brukersPeriode.getTom())
-                .medResultatType(PeriodeResultatType.AVSLÅTT)
-                .medAktiviteter(List.of(aktivitetMedTrekkdager))
-                .build();
-        List<ForeldrepengerUttakPeriode> annenpartPerioder = List.of(annenpartInnvilgetPeriode,
-                annenpartOverlappendeAvslåttMedTrekkdager);
-        var annenpartUttak = new ForeldrepengerUttak(annenpartPerioder);
-        var skalOpprettes = berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat,
-                Optional.of(brukersUttak), Optional.of(annenpartUttak));
-        assertThat(skalOpprettes).isTrue();
-    }
+                .build()))
+            .build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsPeriode.getFom().plusWeeks(1),
+            morsPeriode.getTom().minusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsPeriode));
+        lagreUttak(morBehandling, morUttak);
 
-    // Scenarie 4b - foreldrepenger er innvilget og uttak har negativ saldo
-    @Test
-    public void skal_opprette_berørt_behandling_dersom_revudering_er_innvilget_med_endring_i_uttak_negativ_saldo() {
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.INNVILGET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, false);
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(revurdering), null, new ForeldrepengerGrunnlag());
-        when(uttakInputTjeneste.lagInput(revurdering.getId())).thenReturn(uttakInput);
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
 
-        when(stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(true);
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat, Optional.empty(),
-                Optional.empty())).isTrue();
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isTrue();
     }
 
     @Test
-    public void skal_ikke_opprette_berørt_behandling_dersom_revudering_er_berørt_behandling() {
-        revurdering.getBehandlingÅrsaker()
-                .add(BehandlingÅrsak.builder(BehandlingÅrsakType.BERØRT_BEHANDLING).buildFor(revurdering).get(0));
-        var behandlingsresultat = lagBehandlingsresultat(BehandlingResultatType.FORELDREPENGER_ENDRET,
-                KonsekvensForYtelsen.ENDRING_I_UTTAK, false);
+    public void ikke_berørt_behandling_ved_overlapp_mellom_partene_før_endringsdato() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+        var endringsdato = startDatoMor.plusMonths(3);
 
-        var behandling = revurdering;
-        var uttakInput = new UttakInput(BehandlingReferanse.fra(behandling), null, new ForeldrepengerGrunnlag());
-        when(stønadskontoSaldoTjeneste.erNegativSaldoPåNoenKonto(uttakInput)).thenReturn(false);
-        when(uttakInputTjeneste.lagInput(behandling.getId())).thenReturn(uttakInput);
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(endringsdato));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
 
-        assertThat(berørtBehandlingTjeneste.skalBerørtBehandlingOpprettes(behandlingsresultat,
-                lagUttakResultatPerioder(LocalDate.now().minusDays(12), LocalDate.now().plusDays(2), true, false),
-                lagUttakResultatPerioder(LocalDate.now().plusDays(3), LocalDate.now().plusDays(5), false,
-                        false))).isFalse();
+        lagUttakInput(farBehandling);
+
+        var morsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            endringsdato.minusDays(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var farsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsFørstePeriode.getFom(),
+            morsFørstePeriode.getTom())
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        //Bare fars periode etter endringsdato
+        var farsAndrePerioder = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(endringsdato,
+            endringsdato.plusWeeks(2))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsFørstePeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsFørstePeriode, farsAndrePerioder));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
     }
 
-    private Optional<Behandlingsresultat> lagBehandlingsresultat(BehandlingResultatType behandlingResultatType,
-            KonsekvensForYtelsen konsekvensForYtelsen,
-            boolean harEndretStønadskonto) {
-        return Optional.of(Behandlingsresultat.builder()
-                .medBehandlingResultatType(behandlingResultatType)
-                .leggTilKonsekvensForYtelsen(konsekvensForYtelsen)
-                .medEndretStønadskonto(harEndretStønadskonto)
-                .buildFor(revurdering));
+    @Test
+    public void berørt_behandling_hvis_et_avslag_fører_til_hull_felles_plan() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        //Denne perioden er avslått og skaper derfor hull som må fylles av mor
+        var farsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(
+            morsFørstePeriode.getTom().plusDays(1), morsFørstePeriode.getTom().plusWeeks(1))
+            .medResultatÅrsak(IkkeOppfyltÅrsak.IKKE_STØNADSDAGER_IGJEN)
+            .medAktiviteter(List.of(avslåttUttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morsAndrePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(
+            farsFørstePeriode.getTom().plusDays(1), farsFørstePeriode.getTom().plusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsFørstePeriode, morsAndrePeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsFørstePeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isTrue();
     }
 
-    private Optional<ForeldrepengerUttak> lagUttakResultatPerioder(LocalDate fom,
-            LocalDate tom,
-            boolean medPeriodeForan,
-            boolean medAvslåttPeriode) {
-        List<ForeldrepengerUttakPeriode> perioder = new ArrayList<>();
-        if (medPeriodeForan) {
-            var periode = new ForeldrepengerUttakPeriode.Builder()
-                    .medTidsperiode(fom.minusWeeks(3), fom.minusDays(1))
-                    .medResultatType(PeriodeResultatType.INNVILGET)
-                    .build();
-            perioder.add(periode);
-        }
-        var periode = new ForeldrepengerUttakPeriode.Builder()
-                .medTidsperiode(fom, tom)
-                .medResultatType(PeriodeResultatType.INNVILGET)
-                .build();
-        perioder.add(periode);
-        if (medAvslåttPeriode) {
-            var avslåttPeriode = new ForeldrepengerUttakPeriode.Builder()
-                    .medTidsperiode(tom.plusDays(1), tom.plusDays(7))
-                    .medResultatType(PeriodeResultatType.AVSLÅTT)
-                    .build();
-            perioder.add(avslåttPeriode);
-        }
-        return Optional.of(new ForeldrepengerUttak(perioder, List.of()));
+    @Test
+    public void ikke_berørt_behandling_hvis_bruker_går_tom_for_dager_på_slutten_av_felles_uttaksplan() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var farsFørstePeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(
+            morsFørstePeriode.getTom().plusDays(1), morsFørstePeriode.getTom().plusWeeks(1))
+            .medResultatÅrsak(IkkeOppfyltÅrsak.IKKE_STØNADSDAGER_IGJEN)
+            .medAktiviteter(List.of(avslåttUttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsFørstePeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsFørstePeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
+    }
+
+    @Test
+    public void ikke_berørt_behandling_ved_overlapp_mellom_partene_når_br_konsekvens_er_ingen_endring() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        var brMedIngenEndring = Behandlingsresultat.builderEndreEksisterende(getBehandlingsresultat(farBehandling.getId()))
+            .leggTilKonsekvensForYtelsen(KonsekvensForYtelsen.INGEN_ENDRING)
+            .build();
+        getBehandlingsresultatRepository().lagre(farBehandling.getId(), brMedIngenEndring);
+
+        lagUttakInput(farBehandling);
+
+        var morsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(15))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .medSamtidigUttak(true)
+            .medSamtidigUttaksprosent(new SamtidigUttaksprosent(50))
+            .build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsPeriode.getFom().plusWeeks(1),
+            morsPeriode.getTom().minusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .medSamtidigUttak(true)
+            .medSamtidigUttaksprosent(new SamtidigUttaksprosent(50))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsPeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
+    }
+
+    @Test
+    public void ikke_berørt_behandling_ved_overlapp_mellom_partene_men_behandlingen_er_henlagt() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(15))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsPeriode.getFom().plusWeeks(1),
+            morsPeriode.getTom().minusWeeks(1))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.FEDREKVOTE)))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsPeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
+        var behandlingsresultat = getBehandlingsresultat(farBehandling.getId());
+        var henlagtBehandlingsresultat = Behandlingsresultat.builderEndreEksisterende(behandlingsresultat)
+            .medBehandlingResultatType(BehandlingResultatType.HENLAGT_FEILOPPRETTET)
+            .build();
+        getBehandlingsresultatRepository().lagre(henlagtBehandlingsresultat.getBehandlingId(), henlagtBehandlingsresultat);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
+    }
+
+    @Test
+    public void ikke_berørt_behandling_ved_overlapp_mellom_partene_der_bruker_har_avslått_utsettelse() {
+        var startDatoMor = LocalDate.of(2020, 1, 1);
+
+        var scenario = ScenarioFarSøkerForeldrepenger.forFødsel().medAvklarteUttakDatoer(avklarteDatoer(startDatoMor));
+        var scenarioAnnenpart = ScenarioMorSøkerForeldrepenger.forFødsel();
+        var farBehandling = scenario.lagre(repositoryProvider);
+        var morBehandling = scenarioAnnenpart.lagre(repositoryProvider);
+
+        lagUttakInput(farBehandling);
+
+        var morsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(startDatoMor,
+            startDatoMor.plusWeeks(15))
+            .medResultatÅrsak(InnvilgetÅrsak.KVOTE_ELLER_OVERFØRT_KVOTE)
+            .medAktiviteter(List.of(uttakPeriodeAktivitet(StønadskontoType.MØDREKVOTE)))
+            .build();
+        var farsPeriode = new ForeldrepengerUttakPeriode.Builder().medTidsperiode(morsPeriode.getFom(),
+            morsPeriode.getTom())
+            .medResultatÅrsak(IkkeOppfyltÅrsak.IKKE_STØNADSDAGER_IGJEN)
+            .medResultatType(PeriodeResultatType.AVSLÅTT)
+            .medUtsettelseType(UttakUtsettelseType.ARBEID)
+            .medAktiviteter(List.of(new ForeldrepengerUttakPeriodeAktivitet.Builder().medArbeidsprosent(BigDecimal.ZERO)
+                .medTrekkdager(Trekkdager.ZERO)
+                .medUtbetalingsgrad(Utbetalingsgrad.ZERO)
+                .medAktivitet(new ForeldrepengerUttakAktivitet(UttakArbeidType.FRILANS))
+                .build()))
+            .build();
+        var morUttak = new ForeldrepengerUttak(List.of(morsPeriode));
+        lagreUttak(morBehandling, morUttak);
+
+        var farUttak = new ForeldrepengerUttak(List.of(farsPeriode));
+        lagreUttak(farBehandling, farUttak);
+
+        var resultat = skalBerørtOpprettes(farBehandling, morBehandling);
+        assertThat(resultat).isFalse();
+    }
+
+    private boolean skalBerørtOpprettes(Behandling behandling, Behandling annenpartsBehandling) {
+        return tjeneste.skalBerørtBehandlingOpprettes(getBehandlingsresultat(behandling.getId()), behandling.getId(),
+            annenpartsBehandling.getId());
+    }
+
+    private BehandlingsresultatRepository getBehandlingsresultatRepository() {
+        return repositoryProvider.getBehandlingsresultatRepository();
+    }
+
+    private void lagreUttak(Behandling behandling, ForeldrepengerUttak uttak) {
+        when(foreldrepengerUttakTjeneste.hentUttakHvisEksisterer(behandling.getId())).thenReturn(
+            Optional.ofNullable(uttak));
+    }
+
+    private Behandlingsresultat getBehandlingsresultat(Long behandlingId) {
+        return getBehandlingsresultatRepository().hent(behandlingId);
+    }
+
+    private AvklarteUttakDatoerEntitet avklarteDatoer(LocalDate endringsdato) {
+        return new AvklarteUttakDatoerEntitet.Builder().medOpprinneligEndringsdato(endringsdato).build();
+    }
+
+    private ForeldrepengerUttakPeriodeAktivitet uttakPeriodeAktivitet(StønadskontoType stønadskontoType) {
+        return new ForeldrepengerUttakPeriodeAktivitet.Builder().medArbeidsprosent(BigDecimal.ZERO)
+            .medTrekkdager(new Trekkdager(10))
+            .medTrekkonto(stønadskontoType)
+            .medUtbetalingsgrad(new Utbetalingsgrad(100))
+            .medAktivitet(new ForeldrepengerUttakAktivitet(UttakArbeidType.FRILANS))
+            .build();
+    }
+
+    private ForeldrepengerUttakPeriodeAktivitet avslåttUttakPeriodeAktivitet(StønadskontoType stønadskontoType) {
+        return new ForeldrepengerUttakPeriodeAktivitet.Builder().medArbeidsprosent(BigDecimal.ZERO)
+            .medTrekkdager(Trekkdager.ZERO)
+            .medTrekkonto(stønadskontoType)
+            .medUtbetalingsgrad(Utbetalingsgrad.ZERO)
+            .medAktivitet(new ForeldrepengerUttakAktivitet(UttakArbeidType.FRILANS))
+            .build();
     }
 }
